Quick Config Generation Solution
csv2yml.py will convert any csv file in the local folder to yml, but the Template_Gen.py is currently only grabbing configdata.yml. To use different templates update the template names in 'Template_Gen.py" and update the variable they are keying off of. As shown: access keys off of an IE4K and bb keys off of a C9500.